LPS-85787 Merge FilterParser into FilterProvider for the same reasons.
